+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ master
 ~~~~~~
 
 1) [FEATURE] Add conflict detection for configuration loaders.
+2) [FEATURE] Add a check if a specific configuration item has already been loaded.
 
 5.0.0
 ~~~~~